# swagger/generated-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Telegram Hisobchi Bot API",
    "version": "2.0.0",
    "description": "API documentation for Telegram Hisobchi Bot Backend - Advanced Sales Management System with SellerStock Management",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "TelegramAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-telegram-id",
        "description": "Telegram user ID (REQUIRED for all authenticated endpoints)"
      }
    },
    "schemas": {
...
      "Transfer": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Transfer ID"
          },
          "seller": {
            "type": "string",
            "description": "Seller ID"
          },
          "product": {
            "type": "string",
            "description": "Product ID"
          },
          "quantity": {
            "type": "number",
            "description": "Transferred quantity"
          },
          "type": {
            "type": "string",
            "enum": ["transfer", "return"],
            "description": "Transfer type"
          },
          "status": {
            "type": "string",
            "enum": ["active", "cancelled"],
            "description": "Transfer status"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Transfer creation date"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "TelegramAuth": []
    }
  ],
  "paths": {
...
    "/api/admin/transfers": {
      "get": {
        "tags": ["Admin - Transfer Management"],
        "summary": "Get transfer history",
        "description": "Returns all transfer records with pagination",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Transfers retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transfers": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Transfer"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Admin - Transfer Management"],
        "summary": "Create new transfer(s)",
        "description": "Creates new inventory transfers from warehouse to seller",
        "security": [
          {
            "TelegramAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["sellerId", "items"],
                "properties": {
                  "sellerId": {
                    "type": "string",
                    "description": "Seller ID",
                    "example": "64f1a2b3c4d5e6f7g8h9i0j1"
                  },
                  "items": {
                    "type": "array",
                    "description": "List of products to transfer",
                    "items": {
                      "type": "object",
                      "required": ["productId", "quantity"],
                      "properties": {
                        "productId": {
                          "type": "string",
                          "description": "Product ID",
                          "example": "64f1a2b3c4d5e6f7g8h9i0j2"
                        },
                        "quantity": {
                          "type": "number",
                          "minimum": 1,
                          "description": "Quantity to transfer",
                          "example": 15
                        }
                      }
                    }
                  }
                }
              },
              "example": {
                "sellerId": "64f1a2b3c4d5e6f7g8h9i0j1",
                "items": [
                  {
                    "productId": "64f1a2b3c4d5e6f7g8h9i0j2",
                    "quantity": 15
                  },
                  {
                    "productId": "64f1a2b3c4d5e6f7g8h9i0j3",
                    "quantity": 8
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transfers created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Transfers created successfully"
                    },
                    "transfers": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Transfer"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or insufficient stock",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Omborda yetarli mahsulot yo'q: iPhone 15"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Seller not found"
          }
        }
      }
    },
...
  },
...
}